+++ custom_addons/ak_library_management/models/res_config_settings.py
@@ -1,0 +1,9 @@
+# -*- coding: utf-8 -*-
+
+from odoo import models, fields, api
+
+
+class StockWarehouse(models.TransientModel):
+    _inherit = 'res.config.settings'
+
+    borrow_limit = fields.Integer(string="Borrowing limit", config_parameter="ak_library_management.borrow_limit")